requirements.txt
ruff==0.14.10
requests==2.32.5
python-dotenv==1.2.1
supabase==2.27.2
